move uci from base to own file
